refactor(exposition): extract `Workflow` from `Store`
