+++ extensions/exposition/source/directives/octets/Store.ts
@@ -1,11 +1,10 @@
 import { Readable } from 'node:stream'
-import { posix } from 'node:path'
 import { match } from 'matchacho'
-import { promex } from '@toa.io/generic'
 import { BadRequest, UnsupportedMediaType } from '../../HTTP'
 import { cors } from '../cors'
 import * as schemas from './schemas'
-import type { Maybe } from '@toa.io/types'
+import { Workflow } from './Workflow'
+import type { Unit } from './Workflow'
 import type { Entry } from '@toa.io/extensions.storages'
 import type { Remotes } from '../../Remotes'
 import type { ErrorType } from 'error-value'
@@ -16,11 +15,9 @@ import type { Directive, Input } from './types'
 export class Store implements Directive {
   public readonly targeted = false
 
-  private readonly accept: string | undefined
-  private readonly workflow: Workflow | undefined
+  private readonly accept?: string
+  private readonly workflow?: Workflow
   private readonly discovery: Record<string, Promise<Component>> = {}
-  private readonly remotes: Remotes
-  private readonly components: Record<string, Component> = {}
   private storage: Component | null = null
 
   public constructor
@@ -32,13 +29,15 @@ export class Store implements Directive {
       Array, (types: string[]) => types.join(','),
       undefined)
 
-    this.workflow = match(options?.workflow,
-      Array, (units: Unit[]) => units,
-      Object, (unit: Unit) => [unit],
-      undefined)
+    if (options?.workflow !== undefined) {
+      const units = match<Unit[]>(options.workflow,
+        Array, (units: Unit[]) => units,
+        Object, (unit: Unit) => [unit])
+
+      this.workflow = new Workflow(units, remotes)
+    }
 
     this.discovery.storage = discovery
-    this.remotes = remotes
 
     cors.allowHeader('content-meta')
   }
@@ -65,7 +64,7 @@ export class Store implements Directive {
   private reply (request: Input, storage: string, entry: Entry): Output {
     const body = this.workflow === undefined
       ? entry
-      : Readable.from(this.execute(request, storage, entry))
+      : Readable.from(this.workflow.execute(request, storage, entry))
 
     return { body }
   }
@@ -77,82 +76,6 @@ export class Store implements Directive {
       error)
   }
 
-  /* eslint-disable no-useless-return, max-depth */
-
-  /**
-   * Execute workflow units sequentially, steps within a unit in parallel.
-   * Yield results as soon as they come.
-   *
-   * If you need to change this, it may take a while.
-   */
-  private async * execute (request: Input, storage: string, entry: Entry): AsyncGenerator {
-    yield entry
-
-    const path = posix.join(request.path, entry.id)
-    let interrupted = false
-
-    for (const unit of this.workflow!) {
-      if (interrupted)
-        break
-
-      const steps = Object.keys(unit)
-
-      // unit result promises queue
-      const results = Array.from(steps, promex<unknown>)
-      let next = 0
-
-      // execute steps in parallel
-      for (const step of steps)
-        // these promises are indirectly awaited in the yield loop
-        void (async () => {
-          const endpoint = unit[step]
-          const context: Context = { storage, path, entry }
-          const result = await this.call(endpoint, context)
-
-          if (interrupted)
-            return
-
-          // as a result is received, resolve the next promise from the queue
-          const promise = results[next++]
-
-          if (result instanceof Error) {
-            interrupted = true
-            promise.resolve({ error: { step, ...result } })
-
-            // cancel pending promises
-            results[next].resolve(null)
-          } else
-            promise.resolve({ [step]: result ?? null })
-        })().catch((e) => results[next].reject(e))
-
-      // yield results from the queue as they come
-      for (const promise of results) {
-        const result = await promise
-
-        if (result === null) // canceled promise
-          break
-        else
-          yield result
-      }
-    }
-  }
-
-  private async call (endpoint: string, context: Context): Promise<Maybe<unknown>> {
-    const [operation, component, namespace = 'default'] = endpoint.split('.').reverse()
-    const key = `${namespace}.${component}`
-
-    this.components[key] ??= await this.discover(key, namespace, component)
-
-    return await this.components[key].invoke(operation, { input: context })
-  }
-
-  private async discover (key: string, namespace: string, component: string): Promise<Component> {
-    if (this.discovery[key] === undefined)
-      this.discovery[key] = this.remotes.discover(namespace, component)
-
-    return await this.discovery[key]
-  }
-
   private parseMeta (value: string | string[]): Record<string, string> {
     if (Array.isArray(value))
       value = value.join(',')
@@ -162,27 +85,17 @@ export class Store implements Directive {
     for (const pair of value.split(',')) {
       const eq = pair.indexOf('=')
       const key = (eq === -1 ? pair : pair.slice(0, eq)).trim()
-      const value = eq === -1 ? 'true' : pair.slice(eq + 1).trim()
 
-      meta[key] = value
+      meta[key] = eq === -1 ? 'true' : pair.slice(eq + 1).trim()
     }
 
     return meta
   }
 }
 
-type Unit = Record<string, string>
-type Workflow = Unit[]
-
 export interface Options {
   accept: string | string[]
-  workflow: Workflow | Unit
-}
-
-interface Context {
-  storage: string
-  path: string
-  entry: Entry
+  workflow: Unit[] | Unit
 }
 
 interface StoreInput {
